Renamed the package as well
