+++ worker-workflow-spec/src/main/java/com/github/cafdataprocessing/workflow/spec/WorkflowSpec.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.cafdataprocessing.worker.workflow.shared;
+package com.github.cafdataprocessing.workflow.spec;
 
 import java.util.Objects;
